requirements.txt
flask
torch
torchvision
requests
pillow
scipy
